requirements.txt
Django
djangorestframework
django-cors-headers
channels
redis
psycopg2-binary
python-dotenv
pandas
openpyxl
joblib
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-huggingface
langgraph
faiss-cpu
tiktoken
sentence-transformers
rank-bm25
thefuzz
python-Levenshtein
numpy
fuzzywuzzy
ipykernel